scripts/requirements-docx.txt
# pip install -r scripts/requirements-docx.txt
python-docx>=1.1.0
